refactor: isolate platform inventory loading
